rename some files with dashes
