+++ lib/events/click.js
@@ -1,7 +1,7 @@
 var UBT = require('../kernel');
 var on = require('../lib/on');
 var parents = require('../lib/parents');
-var collectDataFrom = require('../lib/collectdatafrom');
+var collectDataFrom = require('../lib/collect-data-from');
 
 // 监控点击事件
 var key = 'ubt-click';
