+++ lib/events/visit.js
@@ -1,5 +1,5 @@
 var UBT = require('../kernel');
-var collectDataFrom = require('../lib/collectdatafrom');
+var collectDataFrom = require('../lib/collect-data-from');
 
 var key = 'ubt-visit';
 var visitKey = 'ubt-visit-key';
