+++ lib/lib/collect-data-from.js
@@ -1,5 +1,5 @@
-var getRelatedValue = require('./getrelatedvalue');
-var getRelatedMessage = require('./getrelatedmessage');
+var getRelatedValue = require('./get-related-value');
+var getRelatedMessage = require('./get-related-message');
 
 module.exports = function(element) {
