+++ lib/lib/get-related-value.js
